backend-build/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.34.0
google-adk==0.2.0
google-genai==1.9.0
pydantic==2.10.3
python-dotenv==1.1.1
httpx==0.28.1
aiofiles==24.1.0
websockets==13.1
openai==1.51.0
click==8.1.8
rich==13.9.3
pyperclip==1.9.0
PyInstaller==6.11.1
deprecated==1.2.14
